Rename PropertyNameNormalizer.java -> java/de/jflex/ucd_generator/ucd/PropertyNames.java
